PIV#184881401 Linkers should sort by score by default
 If you search by identifier for a string like:

  MS 2010

the linker code turns that into a wildcard query like this:

  MS 2010*

and Solr's query parsing ultimately turns that into a query like this:

  four_part_id:MS OR four_part_id:2010* OR fullrecord:MS OR fullrecord:2010*

(actually there are more fields, but you get the idea)

There are probably *lots* of records matching this criteria, so we're
relying on the scoring to bring the best one to the top.

Since the Search Controller was setting the sort to title by default,
the results from these typeaheads were largely unpredictable and
generally not useful.

This commit sets the linkers to use score sorting by default, and then
modifies the linker to support unescaping the kinds of strings that
Ruby produces.  Turns out they don't agree on whether '+' is an escape
character or not, so hooray.